# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ngx-admin Backend Bundle API",
    "description": "ngx-admin Backend Bundle API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Bearer <my own JWT token>"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User Authentication returning a JWT",
        "description": "Authenticate a user",
        "parameters": [
          {
            "name": "loginDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/loginDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          }
        }
      }
    },
    "/auth/sign-up": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign up new user",
        "description": "Sign up new user",
        "parameters": [
          {
            "name": "signUpDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/signUpDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          }
        }
      }
    },
    "/auth/refresh-token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh user token",
        "description": "Refresh user token",
        "parameters": [
          {
            "name": "refreshTokenDTO",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "refresh_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error message",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Users list",
        "description": "Gets users list",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "filterByfirstName",
            "in": "query",
            "description": "Should filter by first name field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterBylastName",
            "in": "query",
            "description": "Should filter by last name field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterBylogin",
            "in": "query",
            "description": "Should filter by user name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterByemail",
            "in": "query",
            "description": "Should filter by email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterByage",
            "in": "query",
            "description": "Should filter by age",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterBystreet",
            "in": "query",
            "description": "Should filter by street",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterBycity",
            "in": "query",
            "description": "Should filter by city",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterByzipcode",
            "in": "query",
            "description": "Should filter by zip code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Should sort by field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Order direction ASC / DESC",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return list of users (filtered users by query params) and total count of users in response",
            "schema": {
              "type": "object",
              "properties": {
                "items": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "allOf": [
                      {
                        "$ref": "#/definitions/User"
                      }
                    ]
                  }
                },
                "totalCount": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Add new user",
        "description": "Add new user",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "NewUserDto",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by id",
        "description": "Get user by id",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "description": "User Id",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return user by id",
            "schema": {
              "type": "object",
              "allOf": [
                {
                  "$ref": "#/definitions/User"
                }
              ]
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update current user",
        "description": "Update current user",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Module Id",            
            "in": "path"
          },
          {
            "name": "UserDto",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return edited user",
            "schema": {
              "type": "object",
              "allOf": [
                {
                  "$ref": "#/definitions/User"
                },
                {
                  "properties": {
                    "settings": {
                      "type": "object"                      
                    }
                  }
                }
              ]
            }
          },
          "409": {
            "description": "Email is exist",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "metadata": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "field": {
                          "type": "email"
                        },
                        "type": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete user by id",
        "description": "Delete user by id",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "description": "User Id",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return user by id",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "loginDto": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "signUpDto": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      }
    },
...
    "Token": {
      "type": "object",
      "properties": {
        "token": {
          "type": "object",
          "properties": {
            "expires_in": {
              "type": "integer"
            },
            "access_token": {
              "type": "string"
            },
            "refresh_token": {
              "type": "string"
            }
          }
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "age": {
          "type": "integer"
        },
        "login": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "address": {
          "type": "object",
          "properties": {
            "street": {
              "type": "string"
            },
            "city": {
              "type": "string"
            },
            "state": {
              "type": "string"
            },
            "zipCode": {
              "type": "string"
            }
          }
        },
        "profilepic": {
            "type": "string"
        }
      }
    },
...
    "UserDto": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "age": {
          "type": "integer"
        },
        "role": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "address": {
          "type": "object",
          "properties": {
            "street": {
              "type": "string"
            },
            "city": {
              "type": "string"
            },
            "zipCode": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}